P2修复 — count 变量鲁棒性加固
Issue 11 (P2#6+P2#7): count 类型升级 + 临界区保护

P2#7: uint8_t→uint16_t 防 Task4 长时间振荡溢出
- task.h: extern声明 uint8_t→uint16_t
- task.c: 定义同步, 最大65535 (51s→3.6h), 比赛场景永不溢出

P2#6: 3处YawTrack兜底路径 count++/count_debounce 加临界区保护
- __disable_irq() / __enable_irq() 包裹, 消除 ~30ns 指令级竞争窗口
- 仅屏蔽可屏蔽中断, NMI/HardFault 不受影响
- 临界区内仅 count++/count_debounce 两条语句, 开销可忽略 (<10ns)

+++ Drive/task.h
@@ -8,7 +8,7 @@ extern uint16_t selected_task;
 extern volatile uint16_t task_running;     // volatile: ISR 写入（盲开超时停车），主循环读取
 extern uint8_t current_state;
 extern uint8_t last_line_status;
-extern volatile uint8_t count;             // volatile: ISR 与主循环双重递增
+extern volatile uint16_t count;             // volatile: ISR 与主循环双重递增；uint16_t 防 Task4 长时间溢出
 extern uint8_t lap_count;
 extern float current_angle;
 extern volatile uint8_t count_debounce;    // volatile: ISR 与主循环同时写入去抖窗口
